requirements.txt
pandas
apache-airflow-providers-oracle
sqlalchemy
oracledb
apache-airflow-providers-oracle[oracle]